Corrected errors found in canary testing
